requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
torchvision>=0.15.0